styling and moved faq to routes
