+++ app/src/routes/Faq/index.ts
